Move host_user related methods to module
